server/requirements.txt
beautifulsoup4==4.4.1
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
googlemaps==2.4.4
httplib2==0.9.2
Markdown==2.6.6
pytz==2016.4
requests==2.10.0
six==1.10.0
twilio==5.4.0
nanodb==0.4.3